+++ CHANGELOG.md
@@ -24,6 +24,7 @@ The format is based on Keep a Changelog, and this project adheres to Semantic Ve
   - More top padding
 - Blue tagline for ACF
 - Application checklist for ACF
+- Preserve "start" attribute in ols on import
 
 ### Fixed